Fix AttributeError
It seems the columns have string types beforehand and so need to be
explicitly converted to floats

Error:
    AttributeError: 'float' object has no attribute 'sqrt'

    The above exception was the direct cause of the following exception:

    Traceback (most recent call last):
      File "train.py", line 159, in <module>
        training()
      File "train.py", line 49, in __call__
        X_train, y_train, X_val, y_val, X_test, y_test = load_dataset()
      File "/Users/tallamjr/github/tallamjr/origin/astronet/astronet/t2/utils.py", line 331, in load_plasticc
        generated_gp_dataset = __generate_gp_all_objects(object_list, obs_transient)
      File "/Users/tallamjr/github/tallamjr/origin/astronet/astronet/t2/utils.py", line 288, in __generate_gp_all_objects
        gp_predict = fit_2d_gp(df)
      File "/Users/tallamjr/github/tallamjr/origin/astronet/astronet/t2/preprocess.py", line 57, in fit_2d_gp
        obj_flux_error ** 2 + (1e-2 * np.max(obj_flux)) ** 2
      File "/usr/local/anaconda3/envs/astronet/lib/python3.7/site-packages/pandas/core/series.py", line 726, in __array_ufunc__
        result = getattr(ufunc, method)(*inputs, **kwargs)
    TypeError: loop of ufunc does not support argument 0 of type float which has no callable sqrt method

Refs:
    - https://stackoverflow.com/questions/61211476/loop-of-ufunc-does-not-support-argument-0-of-type-float-which-has-no-callable-ex

	modified:   preprocess.py

+++ astronet/t2/preprocess.py
@@ -33,11 +33,9 @@ def fit_2d_gp(obj_data, return_kernel=False, **kwargs):
     """
     guess_length_scale = 20.0  # a parameter of the Matern32Kernel
 
-    #obj_data = preprocess_obs(obj_data, **kwargs)  # preprocess obs
-
-    obj_times = obj_data.mjd
-    obj_flux = obj_data.flux
-    obj_flux_error = obj_data.flux_error
+    obj_times = obj_data.mjd.astype(float)
+    obj_flux = obj_data.flux.astype(float)
+    obj_flux_error = obj_data.flux_error.astype(float)
     obj_wavelengths = obj_data['filter'].map(pb_wavelengths)
 
     def neg_log_like(p):  # Objective function: negative log-likelihood
